+++ src/read.cpp
@@ -124,12 +124,12 @@ RObject read_tokens(List sourceSpec, List tokenizerSpec, ListOf<List> colSpecs,
   size_t n = (n_max < 0) ? 1000 : n_max;
   collectorsResize(collectors, n);
 
-  size_t i = 0, j = 0, cells = 0;
+  size_t i = -1, j = -1, cells = 0;
   for (Token t = tokenizer->nextToken(); t.type() != TOKEN_EOF; t = tokenizer->nextToken()) {
     if (progress && (cells++) % 250000 == 0)
       progressBar.show(tokenizer->progress());
 
-    if (t.col() == 0 && i > 0)
+    if (t.col() == 0 && i != -1)
       checkColumns(&warnings, i, j, p);
 
     if (t.row() >= n) {
@@ -147,16 +147,16 @@ RObject read_tokens(List sourceSpec, List tokenizerSpec, ListOf<List> colSpecs,
     i = t.row();
     j = t.col();
   }
-  checkColumns(&warnings, i, j, p);
+  if (i != -1)
+    checkColumns(&warnings, i, j, p);
 
   if (progress)
     progressBar.show(tokenizer->progress());
   progressBar.stop();
 
-  if (i <= n) {
+  if (i != n - 1) {
     collectorsResize(collectors, i + 1);
   }
-
 
   // Save individual columns into a data frame
   List out(pOut);